+++ src/platforms/test.c
@@ -111,6 +111,15 @@ void adc_gather(void *_adc) {
   (void)_adc;
 }
 
+uint32_t platform_adc_samplerate(void) {
+  return 1000;
+}
+
+uint32_t platform_knock_samplerate(void) {
+  return 1000;
+}
+
+
 void platform_save_config() {}
 
 void platform_load_config() {}